features/requirements.txt
attrs==17.4.0
dask[complete]==0.17.4
flake8==3.5.0
flake8-commas==2.0.0
flake8-import-order==0.17.1
git+https://github.com/has2k1/plotnine.git@7b129e2
git+https://github.com/jdanbrown/potoo.git@18e2f47
joblib==0.11
joypy==0.1.8
jupyter==1.0.0
librosa==0.6.0
pdbpp==0.9.2
pp-ez==0.2.0
pprintpp==0.3.0
pytest==3.5.1
python_speech_features==0.6
pyyaml==3.12
requests-html==0.8.0
rpy2==2.9.2 # >=2.9.2 so we have the fix for fractional -w/-h args to %R
scikit-misc==0.1.1
git+https://github.com/scikit-signal/pytftb@ac2b6a2
git+git://github.com/justinsalamon/skm@c7584ce
statsmodels==0.8.0
structlog==18.1.0

# pydub
pydub==0.21.0
# TODO Switch back to MaxStrange repo after https://github.com/MaxStrange/AudioSegment/pull/2
# git+https://github.com/MaxStrange/AudioSegment.git@86a1214
git+https://github.com/jdanbrown/AudioSegment.git@c8dc862
